Update CallYahooQuotes.apxc
V2 of this Callout. Next Versions will be:

Done:
[+] Update the Related Security record in Salesforce

Todo:
[-] Pass an Array of securities to this class
[-] Expand the Query to Query Any stock from the API that matches the Securities' Symbol from the Array 
[-] Trigger this Automatically when a user navigates to a record page (Input the Stock Name / Symbol from the record that the User is on)

So that... Whenever a User navigates to a specific Security record page in Salesforce, the class is called and the Security Price is updated automatically
OR expand this so that when a User goes to a Household, ALL of the related Securities Records are updated,
then Calculated via a Rollup Summary field / flow / Apex, etc.